Most of GLAreaLight move to Renderer. Fixes #109
