+++ Renderer/Renderer/Lights/AreaLight.h
@@ -5,6 +5,9 @@
 
 #include <Entity/IComponent.h>
 
+#include <GUI/Input/Slider.h>
+#include <GUI/Input/Color.h>
+
 namespace GLEngine::Renderer {
 
 
@@ -14,20 +17,24 @@ public:
 	C_AreaLight(std::shared_ptr<Entity::I_Entity> owner);
 	virtual ~C_AreaLight();
 
+	virtual void DebugDrawGUI() override;
 
 	//================================================================
 	// I_Light
 	[[nodiscard]] virtual Physics::Primitives::C_Frustum GetShadingFrustum() const override;
-	[[nodiscard]] virtual float GetWidth() const { return m_Width; }
-	[[nodiscard]] virtual float GetHeight() const { return m_Height; }
+	[[nodiscard]] float GetWidth() const { return m_WidthSlider.GetValue(); }
+	[[nodiscard]] float GetHeight() const { return m_HeightSlider.GetValue(); }
 
 	virtual Physics::Primitives::S_AABB GetAABB() const override;
 
+
+	[[nodiscard]] glm::vec3 DiffuseColour() const;
+	[[nodiscard]] glm::vec3 SpecularColour() const;
 protected:
 	glm::vec3 m_Normal;
 	glm::vec3 m_UpVector;
-	float			m_Width;
-	float			m_Height;
+	GUI::Input::C_Slider<float> m_WidthSlider, m_HeightSlider;
+	GUI::Input::C_ColorRBG			m_DiffuseColor, m_SpecularColor;
 };
 
 }
